refactor: move sessions commands and tests to cli
