+++ packages/cli/src/commands/sessions/index.ts
@@ -1,9 +1,9 @@
 import color from '@heroku-cli/color'
 import {Command, flags} from '@heroku-cli/command'
-import {CliUx} from '@oclif/core'
+import {ux} from '@oclif/core'
 
-import {OAuthSession} from '../../lib/sessions'
-const sortBy = require('lodash.sortby')
+import {OAuthSession} from '../../lib/sessions/sessions'
+const {sortBy} = require('lodash')
 
 export default class SessionsIndex extends Command {
   static description = 'list your OAuth sessions'
@@ -20,12 +20,12 @@ export default class SessionsIndex extends Command {
     sessions = sortBy(sessions, 'description')
 
     if (flags.json) {
-      CliUx.ux.styledJSON(sessions)
+      ux.styledJSON(sessions)
     } else if (sessions.length === 0) {
-      CliUx.ux.log('No OAuth sessions.')
+      ux.log('No OAuth sessions.')
     } else {
       const printLine: typeof this.log = (...args) => this.log(...args)
-      CliUx.ux.table(sessions, {
+      ux.table(sessions, {
         description: {get: (v: any) => color.green(v.description)},
         id: {},
       }, {'no-header': true, printLine})
