+++ packages/cli/test/unit/commands/sessions/destroy.unit.test.ts
