+++ packages/cli/test/unit/commands/sessions/index.unit.test.ts
